lesson 31, 23.05.2023

+++ ASP.NET/CourseApp/src/webUI/CourseApp.Mvc/Models/PagingInfo.cs
@@ -3,5 +3,5 @@ public class PagingInfo {
     public Int32 TotalItems { get; set; }
     public Int32 ItemsPerPage { get; set; }
     public Int32 CurrentPage { get; set; }
-    public Int32 TotalPage => (Int32)Math.Ceiling((Decimal)this.TotalItems / this.ItemsPerPage);
+    public Int32 TotalPage => (Int32)Math.Ceiling((Decimal)this.TotalItems / (Decimal)this.ItemsPerPage);
 }
